Cleanup larva
