+++ core/src/main/java/org/frankframework/pipes/ForEachChildElementPipe.java
@@ -215,9 +215,6 @@ public class ForEachChildElementPipe extends StringIteratorPipe implements IThre
 
 		private void checkInterrupt() throws SAXException {
 			if (Thread.currentThread().isInterrupted()) {
-				Exception rootException = new InterruptedException("Thread has been interrupted");
-				rootException.fillInStackTrace();
-
 				throw new SAXException("Thread has been interrupted");
 			}
 		}
